Safari-611.1.10/LayoutTests/fast/mediastream/MediaStream-clone-expected.txt


PASS mediaStream.getAudioTracks().length is 1
PASS mediaStream.getVideoTracks().length is 1
newMediaStream = mediaStream.clone()
PASS mediaStream.id is not newMediaStream.id
check tracks
PASS track.id is not newTrack.id
PASS track.kind is newTrack.kind
PASS track.label is newTrack.label
PASS track.readyState is newTrack.readyState
PASS track.enabled is newTrack.enabled
PASS track.muted is newTrack.muted
done
check tracks
PASS track.id is not newTrack.id
PASS track.kind is newTrack.kind
PASS track.label is newTrack.label
PASS track.readyState is newTrack.readyState
PASS track.enabled is newTrack.enabled
PASS track.muted is newTrack.muted
done
PASS successfullyParsed is true

TEST COMPLETE
